requirements.txt
# Neo4j NL2Cypher agent（独立环境，与 llama_factory 隔离）
neo4j>=5.0
pydantic-settings>=2.0
pyahocorasick>=2.0

# Neo4j 可选依赖会 import pandas；与 NumPy 2 对齐，避免 pyarrow 二进制冲突
numpy>=2.0,<3
pandas>=2.2
pyarrow>=16

# LangChain / LangGraph
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-ollama>=0.3.0

# Web API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
